Checkpoint before assistant change: Improve the accuracy of terrain data calculations for more realistic simulations
Corrects type casting in `GeotiffExternalDataReader.java` to avoid potential data loss during terrain calculations.

+++ src/main/java/org/joshsim/geo/external/readers/GeotiffExternalDataReader.java
@@ -174,8 +174,8 @@ public class GeotiffExternalDataReader implements ExternalDataReader {
       }
       
       // Get value from tile
-      int localX = (int)(position.getOrdinate(0) - tileX);
-      int localY = (int)(position.getOrdinate(1) - tileY);
+      int localX = position.getOrdinate(0) - tileX;
+      int localY = position.getOrdinate(1) - tileY;
       double value = tileData[localY][localX];
       
       if (Double.isNaN(value)) {
